+++ install_sdr.sh
@@ -5,34 +5,45 @@ arq=$(uname -m)
 
 apps_common="cmake python3 minimodem swig gr-osmosdr golang"
 apps_osx="libusb sox rtlsdr librtlsdr hackrf airspy inspectrum" # gr-baz gr-fosphor libmirisdr"
-apps_linux_common="build-essential pkg-config libusb-1.0-0 libusb-1.0-0-dev libhackrf0 libhackrf-dev libsoxr0 libsoxr-dev libairspy0 libairspy-dev libmirisdr0 libmidirsdr-dev ibconfig-dev libjpeg-dev libconfig9 libboost-dev sqlite pyqt4-dev-tools liblog4cpp5-dev gnuradio-dev qsstv ax25-apps ax25mail-utils ax25-node ax25-tools ax25-xtools soundmodem libfftw3-dev qt5-default"
+apps_linux_debian_common="build-essential pkg-config libusb-1.0-0 libusb-1.0-0-dev libhackrf0 libhackrf-dev libsoxr0 libsoxr-dev libairspy0 libairspy-dev libmirisdr0 libmidirsdr-dev ibconfig-dev libjpeg-dev libconfig9 libboost-dev sqlite pyqt4-dev-tools liblog4cpp5-dev gnuradio-dev qsstv ax25-apps ax25mail-utils ax25-node ax25-tools ax25-xtools soundmodem libfftw3-dev qt5-default"
 apps_linux_rpi="direwolf "
 apps_linux_ubuntu="librtlsdr-dev libhackrf-dev"
+apps_linux_arch="cmake swig gnuradio gnuradio-osmosdr gnuradio-companion"
 
 pip2="Cheetah lxml matplotlib numpy scipy docutils sphinx pyrtlsdr"
 pip3="urh"
 
 if [ $os == "Linux" ]; then
 
-    # on Debian Linux distributions
-    sudo apt-get update
-    sudo apt-get upgrade
-    sudo apt-get install $apps_common
-    sudo apt-get install $apps_linux_common
+    # DEBIAN LINUX distributions
+    if [ -e /usr/bin/apt ]; then
 
-    # on RaspberryPi
-    if [ $arq == "armv6l" ] || [ $arq == "armv7l" ]; then
-        sudo apt-get install $apps_linux_rpi
-	
-	    cd ~
-        git clone --depth 1 --recursive git@github.com:F5OEO/rpitx.git
-        cd rpitx
-        ./install.sh
-        cd ..
-        # rm -rf rpitx
+        sudo apt-get update
+        sudo apt-get upgrade
+        sudo apt-get install $apps_common
+        sudo apt-get install $apps_linux_debian_common
 
-    else
-        sudo apt-get install $apps_linux_ubuntu
+        # on RaspberryPi
+        if [ $arq == "armv6l" ] || [ $arq == "armv7l" ]; then
+            sudo apt-get install $apps_linux_rpi
+        
+            cd ~
+            git clone --depth 1 --recursive git@github.com:F5OEO/rpitx.git
+            cd rpitx
+            ./install.sh
+            cd ..
+            # rm -rf rpitx
+
+        else
+            sudo apt-get install $apps_linux_ubuntu
+        fi
+
+    # ARCH LINUX distribution
+    elif [ -e /usr/bin/pacman ]; then
+
+        sudo pacman -Sy
+        sudo pacman -S $apps_linux_arch
+
     fi
 
     #   Install OSMCOM drivers
